feat: restore resource ID field on top and refactor Git URL handling in FindingDetail component
